Cleanup code to remove warning messages at buildtime.

+++ src/common/string_utils.cc
@@ -36,7 +36,7 @@ bool EndsWith(const std::string& str, const std::string& sub) {
 std::string ReplaceAll(const std::string& replace,
                        const std::string& from, const std::string& to) {
   std::string str = replace;
-  int pos = str.find(from);
+  size_t pos = str.find(from);
   while (pos != std::string::npos) {
     str.replace(pos, from.length(), to);
     pos = str.find(from, pos+to.length());
